pyproject.toml
[tool.poetry]
name = "vkr_itmo"
version = "0.1.0"
description = ""
authors = ["s.dustov <dustov2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = ">=0.129.0,<0.130.0"
uvicorn = {version = ">=0.44.0,<0.45.0", extras = ["standard"]}
pydantic = ">=2.12.5,<3.0.0"
pydantic-settings = ">=2.13.1,<3.0.0"
sqlalchemy = ">=2.0.46,<3.0.0"
alembic = ">=1.18.4,<2.0.0"
asyncpg = ">=0.31.0,<0.32.0"
psycopg2 = ">=2.9.11,<3.0.0"
python-jose = {version = ">=3.5.0,<4.0.0", extras = ["cryptography"]}
passlib = {version = ">=1.7.4,<2.0.0", extras = ["bcrypt"]}
bcrypt = ">=5.0.0,<6.0.0"
qrcode = {version = ">=8.2,<9.0", extras = ["pil"]}

[tool.poetry.group.dev.dependencies]
mypy = ">=1.19.1,<2.0.0"
pylint = ">=4.0.5,<5.0.0"
ruff = ">=0.15.2,<0.16.0"

[build-system]
requires = ["poetry-core>=1.0.0,<2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.messages_control]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "too-few-public-methods",
]

[tool.pylint.format]
max-line-length = 120

[tool.pylint.basic]
good-names = [
    "id",
    "db",
    "app",
]